app/requirements.txt
appdirs==1.4.4
asgiref==3.5.2
attrs==21.4.0
black==22.3.0
certifi==2022.5.18.1
chardet==4.0.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==3.8.3
flaky==3.7.0
idna==2.10
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
more-itertools==8.13.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==0.13.1
py==1.11.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==3.0.9
pytest==5.4.1
pytest-cov==2.10.1
pytest-django==4.5.2
pytest-mock==3.2.0
pytz==2022.1
regex==2022.6.2
requests==2.25.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
